Two intervals

+++ src/main/java/TwoNumbers.java
@@ -14,15 +14,15 @@ import java.util.Scanner;
 public class TwoNumbers {
                    public static void main(String[] args){
            Scanner input = new Scanner (System.in);
-           int x = input.nextInt();
-           int y = input.nextInt();
+           double x = input.nextInt();
+           double y = input.nextInt();
            /*floor 10 / 6 = 1
                 ceil 10 / 6 = 2
                 round 10 / 6 = 2*/
            
                          System.out.println("floor "+(x)+" / "+(y)+" = "+(int)Math.floor(x/y));
-                       System.out.println("ceil "+(x)+" / "+(y)+" = "+(int)Math.ceil((double)x/y));
-                         System.out.println("round "+(x)+" / "+(y)+" = "+Math.round((float)x/y));
+                       System.out.println("ceil "+(x)+" / "+(y)+" = "+(int)Math.ceil(x/y));
+                         System.out.println("round "+(x)+" / "+(y)+" = "+(int)Math.round(x/y));
            
                    }
 }
